Fun1:
I renamed the variable FilteredLenght to Filtered for conciseness and renamed the variable FilteredParsed to Filtered.
I removed FilteredLength argument in instrument/2 as it is not used in the function.
I removed the value pattern matching in Result as it is not necessary to extract the value.
Fun2:

I removed the pattern matching in Expression and Tokens as they are not necessary.
I removed ParsedLenght and instead return the tuple {Parsed, length(Parsed)}.

Fun3
I changed the condition Statement == 'if' and Statement == 'case' to a case statement.
I removed the unnecessary Filterering and FilteredParsed variables, instead using the result of lists:delete_element/2 directly.
I simplified the calculation of CurrentBranch in the else branch.
I renamed ParsedLenght to Len and CurrentBranch to Curr for conciseness.

Fun4
I renamed Lenght to Len for conciseness.
I removed the unused Holder variable and passed an empty list directly to instrument_parsed/4.

Fun5
I replaced Lenght with Len for conciseness.
I changed the comparison from CurrentBranch =< Lenght to Current > Len for clarity.
I changed the variable name Holder to Acc for clarity.

Fun6
that slightly improves readability and reduces code duplication:

Fun7
by eliminating the creation and subsequent deletion of the tuple Filter and returning InstrumentedClauses directly as the final result:

Fun8 e 9
a more descriptive naming convention and breaking down the code into separate functions to make it easier to understand and maintain.